Bump wasmtime to 0.28.0 (#10)
* Update to latest stable release of wasmtime

Update to wasmtime 0.27.0, this is required to address a security issue
inside of cranelift.

See wasmtime's release notes: https://github.com/bytecodealliance/wasmtime/blob/v0.27.0/RELEASES.md#0270

Signed-off-by: Flavio Castelli <fcastelli@suse.com>

* Bump wasmtime to 0.28.0

Wasmtime went through an important refactor [1] (rendered [2]) in the
0.28.0 version of the project.

The main theme of the refactor was to improve the multi threaded
support, in that the usage of `unsafe` is removed or greatly reduced.

Interior mutability of the `Store` was required before this refactor,
as the host store was shared with a `&T` reference across callbacks
that might require to update the host state. This required interior
mutability and can be error prone.

The `Store` is now uniquely owned, and wrapper types have been added
in order to read or mutate the host store: `AsContext` trait for
reading the host state, and `AsContextMut` trait for reading and
writing the host state.

Mutability is now explicit and requires a `mut store` where
`AsContextMut` trait has to be used.

Support has been added for multiple threads in an easier way that
doesn't involve (or at least, that much) `unsafe` usage. The wasmtime
`Store` type now depends explicitly on the type used to store the host
state, becoming a `Store<T>`. In general, this allows implementors
like us to have a conditionally sendable across threads `Store`,
depending on `T`. If `T` is `Send + Sync`, the `Store<T>` will be
`Send + Sync`, but is not a requirement. Also, this lets us to easily
wrap a not sendable structure on a `Store<Rc<T>>` or even a
`Store<Rc<RefCell<T>>>`.

The way to register and access WASI snapshots has also changed. The
WASI snapshots are not publicly exposed now in the `wasmtime-wasi`
crate. Now, it's possible to let `wasmtime-wasi` register all existing
snapshots in a `Linker` and link to the WASI exports by using the
`get_by_import` function of the linker in order to arrange the imports
when instantiating the WebAssembly module as specified by the
WebAssembly spec [3].

[1] https://github.com/bytecodealliance/rfcs/pull/11
[2] https://github.com/alexcrichton/rfcs-2/blob/new-api/accepted/new-api.md
[3] https://webassembly.github.io/spec/core/exec/modules.html#instantiation

Signed-off-by: Rafael Fernández López <rfernandezlopez@suse.com>

Co-authored-by: Flavio Castelli <fcastelli@suse.com>