wasmtime-provider change Error enum
Updating to wasmtime 0.40 breaks waasmtime-provider own `Error` enum
when the `wasi` feature is disabled.

When the `wasi` feature is disabled, the `wasi_common::Error` is no
longer defined. That causes `wasmtime_provider::Error::WasiError`
variant definition to break.

It is possible to hide this enum variant definition behind a `cfg` check,
and prevent its definition when the `wasi` feature is disabled.
However, when this is done, the code will break in some other parts when
the `?` operator attempts to convert a `anyhow::Error` into a
`wasmtime_provider::Error`.
The `wasmtime` create doesn't define a custom error type, but instead has many
APIs returning a `anyhow::Result` (which is a `Result` returning a
`anyhow::Error` type of error). This includes methods like `wasmtime::Module.new`.

This compilation error can be solved by defining a new variant inside
of `wasmtime_provider::Error`, which transparently maps `anyhow::Error`.
Unfortunately, this breaks the compilation when the `wasi` feature is
enabled. The compilation fails because there are two conflicting
implementations of `std::convert::From<anyhow::Error>`.

This commit removes the `wasmtime_provider::Error::WasiError` variant
and introduces a `Generic` variant that can handle all `anyhow::Error`
transformations.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>