wasmtime provider: allow epoch deadlines
wasmtime has a feature called "epoch deadlines" that makes possible to
abort the wasm code execution after a certain time has elapsed. This
makes possible to prevent endless/slow WASM modules from stalling a waPC
host.

In order to enable this feature more parameters must be provided to the
`wasmtime_provider::WasmtimeEngineProvider`. Instead of growing the list
of parameters used by the different `new_` methods, this commit
introduces a `WasmtimeEngineProviderBuilder` class.

By using the Rust "builder pattern", the creation of the
`WasmtimeEngineProvider` becomes more ergonomic.

The previous `new_` methods are still available, but are currently
deprecated. They can be removed in a future release of the
`wasmtime_provider` create.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>