fix: Improve WasmtimeEngineProvider::clone performance
This commit improves the execution of `WasmtimeEngineProvider::clone()`
by 72%.

This is extremely useful for FaaS-like contexts, where on new requests a
`WasmtimeEngineProvider` instance is cloned just to serve the incoming
data.

This improvement required quite some code refactoring. Details are
below.

== The bottleneck

The bottleneck was inside of the `WasmtimeEngineProvider::init()` method, which is
called by `wapc::WapcHost`.
This method used to take a `wasmtime::Module` and iterate over
a list of all the imported functions.
The code then would build an array with all the functions to be imported,
both the WASI and the waPC ones.
Finally, the code would create a `wasmtime::Instance`.

Instantiating a module is a time consuming operation, that was done
every time `WasmtimeEngineProvider::clone()` was invoked.

This process was needed to register all the waPC functions, which
require to have access to the `host` instance.

This commit moves the `host` instance into the `WapcStore` object, which
is accessible via the `wasmtime::Store`. The store is then accessible
inside of the closures defining the waPC callbacks.

All the waPC callbacks have been rewritten to obtain the `host`
instance by accessing the local store.
This makes possible to register all the waPC callbacs as soon as the
`wasmtime::Linker` is created. This is similar to how WASI functions are made
available to the linker. This allows an important optimization.

When `WasmtimeEngineProvider` is created, the code creates a `wasmtime::Store`
that has inside of it a temporary instance of `WapcHost`.
This instance doesn't have a `host` set, but this isn't a problem since
the functions are not executed yet.
Then, using a `wasmtime::Linker` and this temporary store, we create a
`wasmtime::InstancePre` which is then saved into the
`WasmtimeEngineProvider` instance. Instantiating a module is a time
expensive operation. The `wasmtime::InstancePre` allows us to pay this
cost just once.

When `WasmtimeEngineProvider::init` happes, we create the
final `WapcHost` object using the `host` instance received as
a parameter by the method.
Then we use the `wasmtime::InstancePre::instantiate` method to create
the actual `wasmtime::Instance`.

To summarize: with this code the module instantiation fee is paid only
once, when the first `WasmtimeEngineProvider` is created. All the next
`clone()` operations are not going to be slowed down by it.

== Code cleanups

While rewriting all the waPC callbacks, all the `unwrap` invocations
have been removed. The functions now raise proper `wasmtime::Trap` when
something goes wrong.

== The `WasmtimeEngineProviderPre` struct

With `wasmtime_provider` 1.2.0 release, the
`WasmtimeEngineProviderBuilder` struct was introduced and the `new` methods
of `WasmtimeEngineProvider` got deprecated.

The code inside of the deprecated `new` methods have been moved into
`WasmtimeEngineProviderPre`, which is now consumed by the builder.

The `WasmtimeEngineProviderPre` has a `rehydrate` method that can be
used to generate a `WasmtimeEngineProvider` instance.
This method is faster than `WasmtimeEngineProvider::clone()` by 10
microseconds. I expected it to be faster, that's why I decided to not
make the struct public.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>