update wasmtime (#57)
* fix(wasmtime-provider): build also when wasi feature is disabled

Ensure the code compiles also when the `wasi` feature is disabled.

Note well: some **internal** APIs have been changed. Since these
changes are not visible to the consumers of the library there's
no need to do a major/minor version bump.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>

* chore(deps): update wasmtime

Update the wasmtime-provider crate to consume latest stable release of
wasmtime.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>

* tests: extend GH workflows

Prior to this commit, the GH workflow was running the tests using
`cargo test --workspace`. This however didn't run the wasmtime-provider
test with the `wasi` feature disabled.

Because of that we didn't notice that building without the `wasi`
feature enabled didn't work.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>

---------

Signed-off-by: Flavio Castelli <fcastelli@suse.com>