automation: uniform handling of Rust installation
Prior to this commit, the GH actions installed the rust toolchain using
the now archived `actions-rs/toolchain`.

On top of that, the actions were installing latest `stable` release of
Rust, overriding what was specified inside of our `rust-toolchain.toml`
file. That lead to different behaviour locally and inside of the
runners.
For example, that caused the automated release of `wasmtime-provider`
1.19.0 to fail because of different clippy rules between the local
development and the runners.

GitHub runners have `rustup` installed by default. That, together with
our `rust-toolchain.toml` file is enough to get the right version of
Rust automatically installed.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>