deps: update Rust's WASI target
The old `wasm32-wasi` target has been deprecated and is going to be
completely replaced by `wasm32-wasip1`.

See [this blog post](https://blog.rust-lang.org/2024/04/09/updates-to-rusts-wasi-targets.html).

This commit ensures we use the right WASI target.

As a final note, the `.wasm` file part of the commit is the same program
rebuilt with the new target.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>