multiple fixes (#86)
* chore(deps): update to latest stable release of Rust

Update to Rust 1.79.0. Also, make sure all the required targets are
installed.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>

* automation: uniform handling of Rust installation

Prior to this commit, the GH actions installed the rust toolchain using
the now archived `actions-rs/toolchain`.

On top of that, the actions were installing latest `stable` release of
Rust, overriding what was specified inside of our `rust-toolchain.toml`
file. That lead to different behaviour locally and inside of the
runners.
For example, that caused the automated release of `wasmtime-provider`
1.19.0 to fail because of different clippy rules between the local
development and the runners.

GitHub runners have `rustup` installed by default. That, together with
our `rust-toolchain.toml` file is enough to get the right version of
Rust automatically installed.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>

* deps: update Rust's WASI target

The old `wasm32-wasi` target has been deprecated and is going to be
completely replaced by `wasm32-wasip1`.

See [this blog post](https://blog.rust-lang.org/2024/04/09/updates-to-rusts-wasi-targets.html).

This commit ensures we use the right WASI target.

As a final note, the `.wasm` file part of the commit is the same program
rebuilt with the new target.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>

* fixes: address linter warnings

Address all the linter warnings raised by latest version of clippy

Signed-off-by: Flavio Castelli <fcastelli@suse.com>

---------

Signed-off-by: Flavio Castelli <fcastelli@suse.com>