fix: ensure Go programs built with tinygo >= 0.35.0 run
WASI programs built by tinygo have to be written with a `main` function, even if it's empty.
Starting from tinygo >= 0.35.0, the `main` function calls the WASI process exit function,
which is handled by wasmtime as an Error.

We must check if this error can be converted into a WASI exit
error and, if the exit code is 0, we can ignore it. Otherwise the waPC initialization
will fail.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>