feat!(wasmtime-provider): do not bind epoch deadline to "pre" instances
Do not associate the epoch deadline value to the
WasmtimeEngineProviderPre.

This value can be provided at rehydration time to allow better reuse of
the pre instances.

The use case is: allow the same instance of wapc module to be
instantiated with different epoch deadlines, starting from the same "pre" object.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>